Project structure and code refactor
Closes #12
