+++ src/main/java/com/wintermindset/transaction_service/factory/UserFactory.java
@@ -2,11 +2,11 @@ package com.wintermindset.transaction_service.factory;
 
 import java.time.Instant;
 
-import com.wintermindset.transaction_service.entity.Role;
 import com.wintermindset.transaction_service.entity.UserEntity;
+import com.wintermindset.transaction_service.enums.user.Role;
 
 public interface UserFactory {
-    public UserEntity createUser(
+    UserEntity createUser(
             String username,
             String passwordHash,
             Role role,
